Reorganize to make pip-installable
